Run rtl_433 as a subprocess
I had a lot of trouble getting Ctrl-C to quit the shell pipeline gracefully.

Apparently the semantics of Ctrl-C'ing a shell script are varied and
depend heavily on the exit codes of the subprocesses. In any case, it
wasn't working, and I don't really want to learn more shell to figure
it out. Go's subprocess support seems good, and subprocessing is in line
with other rtl_433 exporters that I've seen.